chore: move blueprint::server to new file
