+++ src/blueprint/from_config.rs
@@ -31,7 +31,7 @@ pub fn config_blueprint(config: &Config) -> Valid<Blueprint> {
   let input_types = config.input_types();
   let schema = to_schema(config)?;
   let definitions = to_definitions(config, output_types, input_types)?;
-  let server: blueprint::Server = Server::try_from(config.server.clone())?;
+  let server: Server = Server::try_from(config.server.clone())?;
   Ok(super::compress::compress(Blueprint { schema, definitions, server }))
 }
 fn to_directive(const_directive: ConstDirective) -> Valid<Directive> {
