+++ src/blueprint/mod.rs
@@ -5,6 +5,9 @@ mod into_schema;
 mod timeout;
 
 // TODO: make it private
+mod server;
 pub mod transform;
+
 pub use blueprint::*;
+pub use server::*;
 pub use timeout::GlobalTimeout;
